amd-sev: Set guest_memfd memory regions in memory_init
